PIRANHA-3221 Error appears when clicking 'Process Now' button in Node.js Debugger several times

+++ site/modules/mdb/static/js/controllers/mdb-detail.js
@@ -152,6 +152,9 @@
                     PopupDialog.error(null, 'You must add the core file.', flushAll);
                     return;
                 }
+                if ($scope.processing) {
+                    return;
+                }
                 flushAll();
                 $scope.processing = true;
                 var callJob = mdb.process({coreFile: $scope.getFilePath(true)}, function (error, job) {
@@ -186,6 +189,9 @@
             };
 
             $scope.cancel = function () {
+                if (!$scope.processing) {
+                    return;
+                }
                 $scope.processing = false;
                 $scope.status = 'Canceling';
                 mdb.cancel($scope.jobId).then(function (status) {
